requirements.txt
scipy==1.5.4
pandas==1.1.4
mne==0.21.1
numpy==1.19.4
h5py==3.1.0
pyEDFlib==0.1.19
requests==2.24.0
scikit_image==0.17.2
xgboost==1.2.1
matplotlib==3.3.2
ray==1.0.0
tqdm==4.51.0
GitPython==3.1.11
scikit_learn==0.23.2
skimage==0.0
